api_faf/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
numpy
streamlit
requests
geopandas
gcsfs
google-genai
google-cloud-bigquery
google-cloud-storage
openrouteservice
openai
python-dotenv
python-multipart
gpxpy
fsspec
folium
